requirements.txt
alembic==1.8.1
astroid==2.12.12
attrs==22.1.0
autopep8==1.7.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
coverage==6.5.0
dill==0.3.5.1
exceptiongroup==1.0.0
Flask==2.2.2
Flask-Alembic==2.0.1
Flask-JWT-Extended==4.4.4
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-MySQL==1.5.2
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
gunicorn==20.1.0
idna==3.4
importlib-metadata==4.12.0
importlib-resources==5.9.0
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
Mako==1.2.3
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-sqlalchemy==0.28.1
mccabe==0.7.0
mock==4.0.3
mysqlclient==2.1.1
packaging==21.3
Pillow==9.3.0
platformdirs==2.5.2
pluggy==1.0.0
pycodestyle==2.9.1
PyJWT==2.5.0
pylint==2.15.4
PyMySQL==1.0.2
pyparsing==3.0.9
pytest==7.2.0
requests==2.28.1
rptree==0.1.1
six==1.16.0
SQLAlchemy==1.4.41
svgwrite==1.4.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.5
Tree==0.2.4
typing-extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
